refactor(storage): extract pure type primitives to packages/storage/src/types.ts
Phase A of the type-safety uplift. Splits the storage package's type
declarations into a foundation file so later phases (batch-API tuple
inference, migrations chain inference, unknown-boundary sweep) have a
canonical location to add typed utilities.

New file `packages/storage/src/types.ts`:

- `StorageArea` / `StorageItemKey` primitives.
- `MetaKey<K>` template-literal helper.
- `KeyParts<K>` \u2014 pure template-literal parse of a StorageItemKey into
  `{ readonly driverArea; readonly driverKey }`. Extracted from the old
  `ResolvedKey<K>` so it can be composed without dragging the
  `WxtStorageDriver` runtime interface along.
- `WatchCallback`, `Unwatch`, `StorageAreaChanges` (`readonly` index
  signature).
- `WxtStorageItemSerializer`, `OnValidationError`,
  `WxtStorageItemOptions`.
- `GetItemOptions`, `RemoveItemOptions`, `SnapshotOptions`
  (`excludeKeys: readonly string[]`).
- `NullablePartial`.

`index.ts` changes:

- Imports the primitives from `./types` and re-exports every public one
  so consumers of `@wxt-dev/storage` see no surface change.
- Replaces the old inline `ResolvedKey<K>` conditional with
  `type ResolvedKey<K extends StorageItemKey> = KeyParts<K> & {\n  driver: WxtStorageDriver }`.
  Stays private because `WxtStorageDriver` is internal.
- Drops the now-unused `Browser` import (`StorageAreaChanges` moved).
- `resolveKey` uses `as unknown as ResolvedKey<K>` for the trust-boundary
  cast \u2014 `KeyParts<K>` has `readonly` fields that TS's overlap check
  treats as non-assignable from the mutable object literal, and the
  double-cast makes the trust boundary explicit at the call site.

Every user-facing type name and export path is preserved. 274 tests
still pass. Oxlint / Publint / TypeScript all clean.

Refs: #1173
